Guidance on Permissions for build.psm1
I do not know much about PowerShell, but wanted to try [building a PR to help test it](https://github.com/PowerShell/PowerShell/pull/17857#issuecomment-1207276196).  (The PR addresses one big reason why I *can't* use PowerShell...[pipe/redirect corruption](https://github.com/PowerShell/PowerShell/issues/1908).)

It is nice that there is a script which will automate the build process.  But the default disposition of powershell is to reject it.  And the link suggested by the warning message given goes to a very long page, with no obvious prescription for the exact incantation to make to run the script.

I don't know if what I did is the "right" answer: (`pwsh -ExecutionPolicy Unrestricted`).  But it worked--so I'm offering it as a first draft.  If there's a better way to advise first-time-builders, then that should be said instead.  But some sort of guidance here would be very helpful.